pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-shield"
version = "0.1.0"
description = "Security proxy for Model Context Protocol (MCP) servers — schema pinning, secret scanning, injection detection"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Lars Weiler", email = "lars@lweiler-lab.dev" },
]
keywords = ["mcp", "security", "proxy", "llm", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "aiohttp>=3.9,<4",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-aiohttp>=1.0",
    "ruff>=0.4",
]

[project.scripts]
mcp-shield = "mcp_shield.cli:main"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
